+++ src/pages/Homepage/HomePage.jsx
@@ -6,7 +6,7 @@ import { useEffect, useState } from "react";
 import Clock from "../../components/Clock/Clock";
 import Button from "../../components/UI/button/Button";
 import desktop from "../../assets/desktop.mp4";
-import Faq from "./Faq";
+import Faq from "../../components/Faq/Faq";
 // import mobile from "../../assets/mobile.mp4"
 
 const HomePage = () => {
